+++ CHANGELOG.rst
@@ -12,6 +12,9 @@ Changelog
 
 **New features**
 
+* Added :meth:`metalearners.metalearner.MetaLearner.fit_all_nuisance` and
+  :meth:`metalearners.metalearner.MetaLearner.fit_all_treatment`.
+
 * Add optional ``store_raw_results`` and ``store_results`` parameters to :class:`metalearners.grid_search.MetaLearnerGridSearch`.
 
 * Renamed :class:`metalearners.grid_search._GSResult` to :class:`metalearners.grid_search.GSResult`.